+++ server/src/test/java/com/inventage/portal/gateway/proxy/middleware/claimToHeader/ClaimToHeaderMiddlewareTest.java
@@ -1,0 +1,98 @@
+package com.inventage.portal.gateway.proxy.middleware.claimToHeader;
+
+import static com.inventage.portal.gateway.proxy.middleware.MiddlewareServerBuilder.portalGateway;
+import static io.vertx.core.http.HttpMethod.GET;
+
+import com.inventage.portal.gateway.proxy.middleware.MiddlewareServer;
+import io.vertx.core.MultiMap;
+import io.vertx.core.Vertx;
+import io.vertx.core.http.HttpHeaders;
+import io.vertx.core.http.RequestOptions;
+import io.vertx.ext.web.RoutingContext;
+import io.vertx.junit5.VertxExtension;
+import io.vertx.junit5.VertxTestContext;
+import java.util.concurrent.atomic.AtomicReference;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+
+@ExtendWith(VertxExtension.class)
+public class ClaimToHeaderMiddlewareTest {
+
+    @Test
+    public void withAuthorizationBearerValue(Vertx vertx, VertxTestContext testCtx) {
+        // given
+        final MultiMap headers = MultiMap.caseInsensitiveMultiMap();
+        headers.add(HttpHeaders.AUTHORIZATION,
+            "Bearer eyJhbGciOiJSUzI1NiIsInR5cCIgOiAiSldUIiwia2lkIiA6ICJ1Y0p4dWNXRDFWZnFSNU56QmtKZng2RnNZYmJHeEcxOHk5bVZrazFYYWJZIn0.eyJleHAiOjE2MTE5Nzg2MjAsImlhdCI6MTYxMTkzNTQyMCwianRpIjoiOWY0MGNiZTEtZDJhMC00N2Q0LWFlNjYtYTI5ZTQyNDg4ZTFmIiwiaXNzIjoiaHR0cDovL2RvY2tlci10ZXN0LmludmVudGFnZS5jb206ODA4MC9hdXRoL3JlYWxtcy9wb3J0YWwiLCJhdWQiOiJhY2NvdW50Iiwic3ViIjoiYWIxNTk0MTMtODkyNC00ZjEwLThkYjEtMTc5NGY0MzFlNzJiIiwidHlwIjoiQmVhcmVyIiwiYXpwIjoiT3JnYW5pc2F0aW9uIiwic2Vzc2lvbl9zdGF0ZSI6IjU3MWEzNTU3LWNiOWMtNDg4ZS05YzMwLWQyZTAwYjQ4MzRkMCIsImFjciI6IjEiLCJyZWFsbV9hY2Nlc3MiOnsicm9sZXMiOlsib2ZmbGluZV9hY2Nlc3MiLCJ1bWFfYXV0aG9yaXphdGlvbiJdfSwicmVzb3VyY2VfYWNjZXNzIjp7Ik9yZ2FuaXNhdGlvbiI6eyJyb2xlcyI6WyJLRVlDTE9BSyJdfSwiYWNjb3VudCI6eyJyb2xlcyI6WyJtYW5hZ2UtYWNjb3VudCIsIm1hbmFnZS1hY2NvdW50LWxpbmtzIiwidmlldy1wcm9maWxlIl19fSwic2NvcGUiOiJwcm9maWxlIGVtYWlsIiwiZW1haWxfdmVyaWZpZWQiOmZhbHNlLCJwcmVmZXJyZWRfdXNlcm5hbWUiOiJzZXJ2aWNlLWFjY291bnQtb3JnYW5pc2F0aW9uIn0.XkgBLTzgUHvmYFu96d55aZdSpkzKEGO6cyQxMfSZt3Ryfp5gLhLywtdXSyaHJ4qgDHDkB3rGcY6NH2tAT4wwDRSCz_4gzOZV5qkwirdvQ8OXwi-Oa16yPjFgj3I8Mn1UO7QkchiiIZ-2J9roVLPhet5McgBSkGB5YyzlWRyFIr2imM5JKuBjT2zVM6HiqzpvRoW11lQBR1h9kbf4sU-RTX4DCHjm8H4pfJx798X0oL6dezirq91QK3gizfR3wboLfEmtA4i-pnTpAOnfOoL9aqKoAweUp83uXJvbPDZjanoJsQhCTiaJprvUtF1CFBWDwtJBWR93Ki9TfBSk3CqTig");
+        final AtomicReference<RoutingContext> routingContext = new AtomicReference<>();
+        MiddlewareServer gateway = portalGateway(vertx, testCtx)
+            .withRoutingContextHolder(routingContext)
+            .withClaimToMiddleware("sub", "X-Uniport-Tenant")
+            .build().start();
+
+        // when
+        gateway.incomingRequest(GET, "/", new RequestOptions().setHeaders(headers), testCtx, (outgoingResponse) -> {
+            // then
+            Assertions.assertEquals("ab159413-8924-4f10-8db1-1794f431e72b", routingContext.get().request().headers().get("X-Uniport-Tenant"));
+            testCtx.completeNow();
+        });
+    }
+
+    @Test
+    public void withAuthorizationBearerNoValue(Vertx vertx, VertxTestContext testCtx) {
+        // given
+        final MultiMap headers = MultiMap.caseInsensitiveMultiMap();
+        final AtomicReference<RoutingContext> routingContext = new AtomicReference<>();
+        MiddlewareServer gateway = portalGateway(vertx, testCtx)
+            .withRoutingContextHolder(routingContext)
+            .withClaimToMiddleware("sub", "X-Uniport-Tenant")
+            .build().start();
+
+        // when
+        gateway.incomingRequest(GET, "/", new RequestOptions().setHeaders(headers), testCtx, (outgoingResponse) -> {
+            // then
+            Assertions.assertFalse(routingContext.get().request().headers().contains("X-Uniport-Tenant"));
+            testCtx.completeNow();
+        });
+    }
+
+    @Test
+    public void withAuthorizationBearerInvalidValue(Vertx vertx, VertxTestContext testCtx) {
+        // given
+        final MultiMap headers = MultiMap.caseInsensitiveMultiMap();
+        headers.add(HttpHeaders.AUTHORIZATION, "Bearer invalidValueForJWT");
+        final AtomicReference<RoutingContext> routingContext = new AtomicReference<>();
+        MiddlewareServer gateway = portalGateway(vertx, testCtx)
+            .withRoutingContextHolder(routingContext)
+            .withClaimToMiddleware("sub", "X-Uniport-Tenant")
+            .build().start();
+
+        // when
+        gateway.incomingRequest(GET, "/", new RequestOptions().setHeaders(headers), testCtx, (outgoingResponse) -> {
+            // then
+            Assertions.assertFalse(routingContext.get().request().headers().contains("X-Uniport-Tenant"));
+            testCtx.completeNow();
+        });
+    }
+
+    @Test
+    public void withAuthorizationBasicValue(Vertx vertx, VertxTestContext testCtx) {
+        // given
+        final MultiMap headers = MultiMap.caseInsensitiveMultiMap();
+        headers.add(HttpHeaders.AUTHORIZATION, "Basic ldUIiwia2lkIiA6ICJ1Y0p4dW");
+        final AtomicReference<RoutingContext> routingContext = new AtomicReference<>();
+        MiddlewareServer gateway = portalGateway(vertx, testCtx)
+            .withRoutingContextHolder(routingContext)
+            .withClaimToMiddleware("sub", "X-Uniport-Tenant")
+            .build().start();
+
+        // when
+        gateway.incomingRequest(GET, "/", new RequestOptions().setHeaders(headers), testCtx, (outgoingResponse) -> {
+            // then
+            Assertions.assertFalse(routingContext.get().request().headers().contains("X-Uniport-Tenant"));
+            testCtx.completeNow();
+        });
+    }
+
+}